Extract reusable CoreData helper methods in a separate file
